Fix - 게시글 리트윗 연동, 쿼리스트링과 LastId 방식으로 인피니트 스크롤 적용

+++ server/routes/post.js
@@ -45,6 +45,10 @@ const upload = multer({
 
 router.post("/", isLoggedIn, upload.none(), async (req, res, next) => {
   try {
+    if (!req.user.id) {
+      return res.status(403).send("로그인이 필요합니다.");
+    }
+
     const { content } = req.body;
     const hashtags = content.match(/#[^\s#]+/g);
 
@@ -106,6 +110,10 @@ router.post("/", isLoggedIn, upload.none(), async (req, res, next) => {
 
 router.post("/:postId/comment", isLoggedIn, async (req, res, next) => {
   try {
+    if (!req.user.id) {
+      return res.status(403).send("로그인이 필요합니다.");
+    }
+
     const post = await Post.findOne({
       where: { id: req.params.postId },
     });
@@ -141,6 +149,10 @@ router.post("/:postId/comment", isLoggedIn, async (req, res, next) => {
 
 router.delete("/:postId", isLoggedIn, async (req, res) => {
   try {
+    if (!req.user.id) {
+      return res.status(403).send("로그인이 필요합니다.");
+    }
+
     Post.destroy({
       where: { id: req.params.postId, UserId: req.user.id },
     });
@@ -155,6 +167,10 @@ router.delete("/:postId", isLoggedIn, async (req, res) => {
 /* LIKE */
 router.patch("/:postId/like", isLoggedIn, async (req, res, next) => {
   try {
+    if (!req.user.id) {
+      return res.status(403).send("로그인이 필요합니다.");
+    }
+
     const id = parseInt(req.params.postId);
     const post = await Post.findOne({ where: { id } });
 
@@ -174,6 +190,10 @@ router.patch("/:postId/like", isLoggedIn, async (req, res, next) => {
 /* UN LIKE */
 router.delete("/:postId/like", isLoggedIn, async (req, res, next) => {
   try {
+    if (!req.user.id) {
+      return res.status(403).send("로그인이 필요합니다.");
+    }
+
     const id = parseInt(req.params.postId);
     const post = await Post.findOne({ where: { id } });
 
@@ -192,9 +212,111 @@ router.delete("/:postId/like", isLoggedIn, async (req, res, next) => {
 
 /* IMAGES UPLOAD */
 router.post("/images", isLoggedIn, upload.array("image"), (req, res, next) => {
+  if (!req.user.id) {
+    return res.status(403).send("로그인이 필요합니다.");
+  }
+
   // 이미지 업로드 후
   console.log(req.files);
   res.json(req.files.map((file) => file.filename));
 }); // POST /post/images
+
+/* RETWEET */
+router.post("/:postId/retweet", isLoggedIn, async (req, res, next) => {
+  try {
+    if (!req.user.id) {
+      return res.status(403).send("로그인이 필요합니다.");
+    }
+
+    const post = await Post.findOne({
+      where: { id: parseInt(req.params.postId) },
+      include: [
+        {
+          model: Post,
+          as: "Retweet",
+        },
+      ],
+    });
+
+    if (!post) {
+      return res.status(403).send("존재하지 않는 게시글잆니다.");
+    }
+
+    if (
+      req.user.id === post.UserId ||
+      (post.Retweet && post.Retweet.UserId === req.user.id)
+    ) {
+      /* 
+        - 자신의 게시글 Retweet 할 수 없도록 체크
+        - 남이 자신의 게시글 Retweet 한 것을 다시 Retweet 할 수 없도록 체크
+      */
+      return res.status(403).send("자신의 글은 리트윗할 수 없습니다.");
+    }
+
+    const retweetTargetId = post.RetweetId || post.id;
+    const exPost = await Post.findOne({
+      where: {
+        UserId: req.user.id,
+        RetweetId: retweetTargetId,
+      },
+    });
+
+    if (exPost) {
+      return res.status(403).sned("이미 리트윗한 게시글입니다.");
+    }
+
+    const retweet = await Post.create({
+      UserId: req.user.id,
+      RetweetId: retweetTargetId,
+      content: "retweet",
+    });
+
+    const retweetWithPrevPost = await Post.findOne({
+      where: { id: retweet.id },
+      include: [
+        {
+          model: Post,
+          as: "Retweet",
+          include: [
+            {
+              model: User,
+              attributes: ["id", "nickname"],
+            },
+            { model: User, attributes: ["id"], as: "Likers" },
+            {
+              model: Image,
+            },
+          ],
+        },
+        {
+          model: User,
+          attributes: ["id", "nickname"],
+        },
+        {
+          model: Image,
+        },
+        {
+          model: Comment,
+          include: [
+            {
+              model: User,
+              attributes: ["id", "nickname"],
+            },
+          ],
+        },
+        // {
+        //   model: User, // 좋아요 누른 사람
+        //   as: "Likers",
+        //   attributes: ["id"],
+        // },
+      ],
+    });
+
+    res.status(201).json(retweetWithPrevPost);
+  } catch (error) {
+    console.error(error);
+    next(error);
+  }
+}); // POST /post/1/retweet
 
 module.exports = router;
